feat: complete migration from hash router to Next.js App Router
- Replace hash-based routing with proper Next.js App Router pages
- Create route structure: /project/[id]/session/[id]/agent/[id]
- Extract all logic from LaceApp.tsx into route-specific components
- Remove hash router and custom state management system
- Fix SSE EventSource lifecycle management for page navigation
- Implement automatic cascade clearing through route structure
- Add useURLState hook to replace useHashRouter functionality

Navigation now works seamlessly: Home → Project → Session → Agent
Real-time SSE streaming works correctly across all page transitions
