+++ packages/web/components/providers/AgentProvider.tsx
@@ -13,7 +13,6 @@ import React, {
   type ReactNode,
 } from 'react';
 import { useAgentManagement } from '@/hooks/useAgentManagement';
-import { useHashRouter } from '@/hooks/useHashRouter';
 import type { SessionInfo, AgentInfo, ThreadId } from '@/types/core';
 import type { CreateAgentRequest } from '@/types/api';
 
@@ -53,10 +52,13 @@ const AgentContext = createContext<AgentContextType | null>(null);
 interface AgentProviderProps {
   children: ReactNode;
   sessionId: string | null;
+  selectedAgentId?: string | null; // Agent ID from URL params
   onAgentChange?: (agentId: string | null) => void;
 }
 
-export function AgentProvider({ children, sessionId, onAgentChange }: AgentProviderProps) {
+export function AgentProvider({ children, sessionId, selectedAgentId, onAgentChange }: AgentProviderProps) {
+  console.warn('[AGENT_PROVIDER] Initialized with sessionId:', sessionId);
+  
   // Get agent data from pure data hook
   const {
     sessionDetails,
@@ -68,11 +70,8 @@ export function AgentProvider({ children, sessionId, onAgentChange }: AgentProvi
     updateAgent,
   } = useAgentManagement(sessionId);
 
-  // Get selection state from hash router
-  const { agent: selectedAgent, setAgent: setSelectedAgent } = useHashRouter();
-
-  // Track whether auto-selection has been performed for this session
-  const hasAutoSelectedRef = useRef<string | null>(null);
+  // Use agent from URL params, not hash router
+  const selectedAgent = selectedAgentId;
 
   // Compute derived state based on data + selection
   const foundAgent = useMemo(() => {
@@ -102,12 +101,11 @@ export function AgentProvider({ children, sessionId, onAgentChange }: AgentProvi
   // Selection actions
   const selectAgent = useCallback(
     (agentId: ThreadId | string | null) => {
-      setSelectedAgent(agentId);
       if (onAgentChange) {
         onAgentChange(agentId);
       }
     },
-    [setSelectedAgent, onAgentChange]
+    [onAgentChange]
   );
 
   const onAgentSelect = useCallback(
@@ -118,28 +116,6 @@ export function AgentProvider({ children, sessionId, onAgentChange }: AgentProvi
     },
     [selectAgent]
   );
-
-  // Auto-select coordinator agent when entering a session (only once per session)
-  useEffect(() => {
-    if (sessionId && sessionDetails && !selectedAgent && hasAutoSelectedRef.current !== sessionId) {
-      // Coordinator agent has the same threadId as the sessionId
-      const coordinatorAgent = sessionDetails.agents?.find((agent) => agent.threadId === sessionId);
-
-      if (coordinatorAgent) {
-        selectAgent(coordinatorAgent.threadId);
-        hasAutoSelectedRef.current = sessionId;
-      }
-    }
-
-    // Reset auto-selection tracking when session changes
-    if (
-      sessionId !== hasAutoSelectedRef.current &&
-      hasAutoSelectedRef.current !== null &&
-      !selectedAgent
-    ) {
-      hasAutoSelectedRef.current = null;
-    }
-  }, [sessionId, sessionDetails, selectedAgent, selectAgent]);
 
   const value: AgentContextType = useMemo(
     () => ({
